+++ changelog.md
@@ -1,14 +1,35 @@
 # Changelog
 
-**Test Status**
-* Storage API tests passed: 23 / 28
+**./fox-api/database-info.js - Renamed Functions**
+* 'getDatabaseInformation' to 'getFolderContents'
+	* Called publicly as 'getContents'
+* 'readDatabaseFolder' to 'readTargetFolder'
 
 ---
 
-**./routes/storage.js**
-* New file - Defines Storage API endpoints.
+**./fox-api/database-info.js - getFolderContents**
+* Renamed parameters:
+	* 'rootPathString' to 'folderPathString'
+	* 'dbInfoCallback' to 'foldContsCallback'
+* Renamed variables:
+	* 'databaseFolderExists' to 'targetFolderExists'
+	* 'databaseArrayObject' to 'entryArrayObject'
 
 ---
 
-**./server.js**
-* Added Storage API endpoints.
+**./fox-api/database-info.js - readTargetFolder**
+* Renamed parameters:
+	* 'rootPathStr' to 'folderPathStr'
+	* 'databaseArrayObj' to 'entryArrayObj'
+
+---
+
+**./fox-api/database-info.js - loopFolderContents**
+* Renamed parameters:
+	* 'rootPath' to 'folderPath'
+	* 'databaseArray' to 'entryArray'
+
+---
+
+**./service.main.js getDatabaseSize**
+* Changed 'databaseInfo.getDatabases' to 'databaseInfo.getContents'